Fixed "3: couldn't match all kex parts" error
After many false starts and banging my had against a wall for a while I
found that; the "3: couldn't match all kex parts" error (at least in our
case) was being generated because we were setting the keys as strings
but it was expecting them as bytes.

I also found that conch ssh can't handle PKCS8 formatted keys. But it
seems happy with TraditionalOpenSSL (a.k.a PKCS1)